Update borrowed nav bar to get borrowed books
Remove the test borrow books button and the function
to make a GET request to the server for borrowed books
in Book.js. Shift the function to make the GET request
into TopNavBar.js, where the function will be called
when the borrowed nav bar is clicked.
