+++ src/Book.js
@@ -31,27 +31,6 @@ function Book(props) {
         }
     }
 
-    const getBorrowedBooks = async () => {
-        // test function to get borrowed books of currently logged in user.
-        // makes a GET request to server.
-        console.log(`getting borrowed books of: ${currentUser.username}`)
-        try {
-            const response = await fetch(`http://localhost:5000/borrowed-books/${currentUser.username}`, {
-                method: 'GET',
-            })
-            if (!response.ok) {
-                throw new Error(`HTTP error! Status: ${response.status}`);
-            }
-            const data = await response.json();
-            console.log(`Borrowed books: ${data.borrowed_books}`);
-            data.borrowed_books.forEach(book => {
-                console.log(book.title);
-            });
-        } catch (error) {
-            console.error('Error fetching borrowed books:', error);
-        }
-    }
-
     const listItems = props.books.map(book => <div className="books-card-display" key={book.id}>
         <p><img src={book.image} alt="cover page of book" /></p>
         <p>{book.title}</p>
@@ -65,7 +44,6 @@ function Book(props) {
         <div style={{display:'flex', marginTop:'5px'}}>
             <button id="borrow-button" onClick={() => handleBorrow(book)} style={{marginTop:"10px", fontFamily: "Georgia, 'Times New Roman', Times, serif"}}>Borrow</button>
             <button id="wishlist-button" style={{marginTop:"10px", fontFamily: "Georgia, 'Times New Roman', Times, serif"}}>Add to Wishlist</button>
-            <button onClick={getBorrowedBooks}>get borrowed books</button>
         </div>
     </div>)
